Turn tests into integration tests
