+++ crates/oro-pack/src/lib.rs
@@ -82,46 +82,3 @@ impl OroPack {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn paths_no_files_field() {
-        let mut cwd = env::current_dir().unwrap();
-        cwd.push("fixtures/implicit_files");
-        env::set_current_dir(cwd).unwrap();
-
-        let mut pack = OroPack::new();
-
-        let expected_paths = vec![
-            Path::new("package.json"),
-            Path::new("src/index.js"),
-            Path::new("src/module.js"),
-        ];
-
-        pack.load_package_json();
-
-        let files = pack.get_pkg_files();
-
-        assert_eq!(expected_paths, files);
-    }
-
-    #[test]
-    fn paths_respect_files() {
-        let mut cwd = env::current_dir().unwrap();
-        cwd.push("fixtures/explicit_files");
-        env::set_current_dir(cwd).unwrap();
-
-        let mut pack = OroPack::new();
-
-        pack.load_package_json();
-
-        let expected_paths = vec![Path::new("src/module.js")];
-
-        let pkg_files = pack.get_pkg_files();
-
-        assert_eq!(expected_paths, pkg_files);
-    }
-}
