+++ crates/oro-pack/tests/explicit.rs
@@ -1,0 +1,20 @@
+use oro_pack::*;
+use std::env;
+use std::path::Path;
+
+#[test]
+fn paths_respect_files() {
+    let mut cwd = env::current_dir().unwrap();
+    cwd.push("fixtures/explicit_files");
+    env::set_current_dir(cwd).unwrap();
+
+    let mut pack = OroPack::new();
+
+    pack.load_package_json();
+
+    let expected_paths = vec![Path::new("src/module.js")];
+
+    let pkg_files = pack.get_pkg_files();
+
+    assert_eq!(expected_paths, pkg_files);
+}
